ml_service/requirements.txt
torch
transformers
pika
psycopg2-binary
redis